Simplify VFS mount logging to reduce kernel output
Reduced the verbose mount logging for tmpfs directories to single-line
messages. This makes the boot log cleaner and helps identify issues.

Current status: Kernel successfully mounts all tmpfs directories and
loads init. Init runs and begins processing /etc/fstab but hangs during
a write syscall while mounting sysfs.

Next: Debug write syscall hang during fstab processing.

+++ crates/syscall/syscall/src/lib.rs
@@ -582,7 +582,7 @@ pub fn dispatch(
             arg3,
             arg4 as usize,
             arg5,
-            arg6 as usize,
+            arg6,
         ),
         nr::UMOUNT => vfs::sys_umount(arg1, arg2 as usize, arg3 as u32),
